pyproject.toml
[project]
name = "subtitle-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<=3.13"
dependencies = [

    "paddleocr>=3.0.0",
    "opencv-python>=4.5.0,<4.11",
    "numpy==1.26.4",
]

# uv pip install paddlepaddle-gpu==3.0.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/
# python -m pip install paddlepaddle-gpu==3.2.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu126/

[[tool.uv.index]]
 name = "paddlepaddle-gpu"
 url = "https://www.paddlepaddle.org.cn/packages/stable/cu118/"
 explicit = true

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
explicit = true

[tool.uv.pip]
index-url = "https://mirrors.aliyun.com/pypi/simple/"